Added logout functionality and more styling

+++ src/app/component/pop-up/pop-up.component.ts
@@ -1,9 +1,9 @@
 import { Component, Inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
-import {
-  MAT_DIALOG_DATA, MatDialogModule
-} from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
 import { ResumeService } from '../../services/resume.service';
+import { AuthService } from '../../services/auth.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-pop-up',
@@ -16,18 +16,23 @@ export class PopUpComponent {
   id?: string;
   message?: string;
   header?: string;
+  action?: string;
 
-  constructor(@Inject(MAT_DIALOG_DATA) public data: any, private resumeService: ResumeService) {
+  constructor(@Inject(MAT_DIALOG_DATA) public data: any, private resumeService: ResumeService, private authService: AuthService, private router: Router) {
     this.id = data.id;
     this.message = data.message;
     this.header = data.header;
+    this.action = data.action;
   }
 
-  delete() {
-    if (this.id) {
+  executeAction() {
+    if (this.id && this.action === 'delete') {
       this.resumeService.deleteResume(this.id);
-    } else {
-      console.error('No id provided for deletion');
+    };
+    if (this.action === 'logout') {
+      console.log('logout');
+      this.authService.logout();
+      this.router.navigate(['/login']);
     }
   }
